refactor: move UsersView into separate component on Admins page
