+++ src/pages/AdminPanel/AdminPanel.styles.tsx
@@ -67,30 +67,6 @@ export const useStyle = makeStyles((theme: Theme) => {
         fill: COLOR_BLACK_5,
       },
     },
-    searchContainer: {
-      ...flexHelper(),
-      marginTop: theme.spacing(3),
-    },
-    searchContainerField: {
-      marginRight: theme.spacing(2),
-    },
-    collapsibleList: {
-      marginTop: theme.spacing(1.5),
-    },
-    pagination: {
-      '& li': {
-        '& > button': {
-          border: `1px solid ${theme.palette.type === 'dark' ? COLOR_BLACK_1 : COLOR_GREY_2}`,
-          borderRadius: theme.spacing(1),
-          background: `${theme.palette.type === 'dark' ? COLOR_BLACK_1 : COLOR_GREY_3}`,
-        },
-        '&:first-child > button, &:last-child > button': {
-          height: 44,
-          width: 44,
-          border: `1px solid ${theme.palette.type === 'dark' ? COLOR_BLACK_3 : COLOR_GREY_3}`,
-        },
-      },
-    },
 
     collapseBox: {
       padding: '32px 0',
@@ -101,9 +77,6 @@ export const useStyle = makeStyles((theme: Theme) => {
     rowText: {
       fontSize: '14px',
       lineHeight: '24px',
-    },
-    adminsSwitch: {
-      ...flexHelper('flex-start'),
     },
   });
 });
